Move subscriptions related logic to new Subscriptions namespace
