+++ spec/factories/referral_code.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 FactoryBot.define do
-  factory :referral_code
+  factory :referral_code, class: 'Subscriptions::ReferralCode'
 end
