+++ spec/factories/subscription_month.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 FactoryBot.define do
-  factory :subscription_month do
+  factory :subscription_month, class: 'Subscriptions::SubscriptionMonth' do
     start_at { Time.now }
     co2e { 1000 }
     user
